# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Smart Job Portal with ATS Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  job_seeker: "job_seeker",
  recruiter: "recruiter",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  job_seeker: "job_seeker",
  recruiter: "recruiter",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  full_time: "full_time",
  part_time: "part_time",
  contract: "contract",
  remote: "remote",
} as const;

export interface Job {
  id: number;
  title: string;
  company: string;
  description: string;
  skills: string[];
  location: string;
  type: JobType;
  /** @nullable */
  salaryMin: number | null;
  /** @nullable */
  salaryMax: number | null;
  recruiterId: number;
  /** @nullable */
  recruiterName: string | null;
  applicationCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateJobBodyType =
  (typeof CreateJobBodyType)[keyof typeof CreateJobBodyType];

export const CreateJobBodyType = {
  full_time: "full_time",
  part_time: "part_time",
  contract: "contract",
  remote: "remote",
} as const;

export interface CreateJobBody {
  title: string;
  company: string;
  description: string;
  skills: string[];
  location: string;
  type: CreateJobBodyType;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
}

export type UpdateJobBodyType =
  (typeof UpdateJobBodyType)[keyof typeof UpdateJobBodyType];

export const UpdateJobBodyType = {
  full_time: "full_time",
  part_time: "part_time",
  contract: "contract",
  remote: "remote",
} as const;

export interface UpdateJobBody {
  title?: string;
  company?: string;
  description?: string;
  skills?: string[];
  location?: string;
  type?: UpdateJobBodyType;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewing: "reviewing",
  shortlisted: "shortlisted",
  rejected: "rejected",
  accepted: "accepted",
} as const;

export interface Application {
  id: number;
  userId: number;
  jobId: number;
  status: ApplicationStatus;
  /** @nullable */
  coverLetter: string | null;
  /** @nullable */
  atsScore: number | null;
  /** @nullable */
  jobTitle: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  applicantName: string | null;
  /** @nullable */
  applicantEmail: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateApplicationBody {
  jobId: number;
  coverLetter?: string;
  /** @nullable */
  resumeId?: number | null;
}

export type UpdateApplicationStatusBodyStatus =
  (typeof UpdateApplicationStatusBodyStatus)[keyof typeof UpdateApplicationStatusBodyStatus];

export const UpdateApplicationStatusBodyStatus = {
  pending: "pending",
  reviewing: "reviewing",
  shortlisted: "shortlisted",
  rejected: "rejected",
  accepted: "accepted",
} as const;

export interface UpdateApplicationStatusBody {
  status: UpdateApplicationStatusBodyStatus;
}

export interface Resume {
  id: number;
  userId: number;
  fileName: string;
  extractedText: string;
  skills: string[];
  createdAt: string;
}

export interface UploadResumeBody {
  fileName: string;
  extractedText: string;
  skills: string[];
}

export interface AtsScore {
  id: number;
  resumeId: number;
  /** @nullable */
  jobId: number | null;
  score: number;
  matchedKeywords: string[];
  missingKeywords: string[];
  totalKeywords: number;
  /** @nullable */
  jobTitle: string | null;
  createdAt: string;
}

export interface AnalyzeResumeBody {
  resumeId: number;
  /** @nullable */
  jobId?: number | null;
  jobDescription: string;
  jobTitle: string;
}

export interface SeekerDashboard {
  totalApplications: number;
  pendingApplications: number;
  shortlisted: number;
  rejected: number;
  accepted: number;
  avgAtsScore: number;
  recentApplications: Application[];
  topAtsScores: AtsScore[];
}

export interface RecruiterDashboard {
  totalJobs: number;
  activeJobs: number;
  totalApplications: number;
  pendingReview: number;
  shortlisted: number;
  recentApplications: Application[];
  jobsWithCounts: Job[];
}

export type PlatformStatsJobsByType = {
  full_time: number;
  part_time: number;
  contract: number;
  remote: number;
};

export interface PlatformStats {
  totalJobs: number;
  totalUsers: number;
  totalApplications: number;
  totalResumes: number;
  avgAtsScore: number;
  jobsByType: PlatformStatsJobsByType;
}

export type ListJobsParams = {
  search?: string;
  location?: string;
  type?: string;
};

export type ListApplicationsParams = {
  jobId?: number;
  userId?: number;
};

export type ListAtsScoresParams = {
  resumeId?: number;
};