# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 3D Smart Job Portal with ATS Resume Analyzer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (filtered by user or job)
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ats/analyze:
    post:
      operationId: analyzeResume
      tags: [ats]
      summary: Analyze resume against a job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumeBody"
      responses:
        "200":
          description: ATS analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtsScore"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ats/scores:
    get:
      operationId: listAtsScores
      tags: [ats]
      summary: List ATS scores for a resume
      parameters:
        - name: resumeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ATS scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtsScore"
  /ats/scores/{id}:
    get:
      operationId: getAtsScore
      tags: [ats]
      summary: Get ATS score by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ATS score details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtsScore"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [job_seeker, recruiter, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [job_seeker, recruiter]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
...
    Application:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [pending, reviewing, shortlisted, rejected, accepted]
        coverLetter:
          type: ["string", "null"]
        atsScore:
          type: ["number", "null"]
        jobTitle:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        applicantName:
          type: ["string", "null"]
        applicantEmail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - jobId
        - status
        - coverLetter
        - atsScore
        - jobTitle
        - company
        - applicantName
        - applicantEmail
        - createdAt
        - updatedAt
    CreateApplicationBody:
      type: object
      properties:
        jobId:
          type: integer
        coverLetter:
          type: string
        resumeId:
          type: ["integer", "null"]
      required:
        - jobId
    UpdateApplicationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewing, shortlisted, rejected, accepted]
      required:
        - status
...
    AtsScore:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        jobId:
          type: ["integer", "null"]
        score:
          type: number
        matchedKeywords:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
        totalKeywords:
          type: integer
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - resumeId
        - jobId
        - score
        - matchedKeywords
        - missingKeywords
        - totalKeywords
        - jobTitle
        - createdAt
    AnalyzeResumeBody:
      type: object
      properties:
        resumeId:
          type: integer
        jobId:
          type: ["integer", "null"]
        jobDescription:
          type: string
        jobTitle:
          type: string
      required:
        - resumeId
        - jobDescription
        - jobTitle
...