# lib/api-zod/src/generated/types/updateApplicationStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Smart Job Portal with ATS Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateApplicationStatusBodyStatus } from "./updateApplicationStatusBodyStatus";

export interface UpdateApplicationStatusBody {
  status: UpdateApplicationStatusBodyStatus;
}